--- app/components/my-lists.js
@@ -1,4 +1,0 @@
-import Ember from 'ember';
-
-export default Ember.Component.extend({
-});
