--- app/models/lists.js
@@ -1,6 +1,0 @@
-import DS from 'ember-data';
-
-export default DS.Model.extend({
-  title:DS.attr('string'),
-  list: DS.hasMany('item', { async: false, inverse: null })
-});
