--- app/routes/lists.js
@@ -1,9 +1,0 @@
-import Ember from 'ember';
-
-export default Ember.Route.extend({
-  model(params) {
-    return this.store.findRecord('list', params.list_id).then(function(list) {
-      return list.map(function(x) { return x.toJSON(); });
-    });
-  }
-});
