--- app/templates/components/my-lists.hbs
@@ -1,9 +1,0 @@
-<h2>{{title}}</h2>
-
-<ul>
-  {{#each list as |list|}}
-    <li>{{list}}</li>
-  {{/each}}
-</ul>
-
-{{yield}}
