+++ app/templates/list.hbs
@@ -1,4 +1,3 @@
-
 <center>
 
 <h3><b>whee list</b></h3>
